Create maxIceCreamsBars.cpp
This is a function that returns the maximum number of ice cream cones that can be bought with a given amount of coins. It does so by first sorting the costs of the ice cream cones in ascending order. It then iterates through the costs of the ice cream cones and checks if the cost of the current ice cream cone is greater than the available coins. If it is, then the function breaks out of the loop because it is not possible to buy any more ice cream cones. If the cost of the current ice cream cone is less than or equal to the available coins, then the function subtracts the cost of the current ice cream cone from the available coins and increments the number of ice cream cones bought. Finally, the function returns the number of ice cream cones bought.