Create gasStation.cpp
The canCompleteCircuit function takes in two vectors, gas and cost, representing the amount of gas available at each gas station and the cost to travel from that station to the next station in the circular route.

The function first initializes three variables: total_gas and total_cost, which will be used to store the total amount of gas and cost for the entire route; and curr_gas, which will be used to store the current amount of gas in the car's tank.

It also initializes a variable called starting_point which will be used to store the index of the starting gas station for a potential route.

The function then iterates through each gas station. For each station, it adds the amount of gas and cost at that station to the total_gas and total_cost variables. It then adds the amount of gas gained at that station to the curr_gas variable and subtracts the cost of traveling to the next station.

If the curr_gas variable becomes negative at any point, it means that the car does not have enough gas to reach the next station. In this case, the function resets the starting_point to the next gas station and resets the curr_gas variable to 0.

Finally, the function checks whether the total gas is less than the total cost. If it is, it returns -1, indicating that it is not possible to complete the circular route. Otherwise, it returns the starting_point, which is the index of the gas station at which the circular route can be started.