Create preorderBinaryTreeTraversal.cpp
This code defines a solution to the problem of performing a preorder traversal of a binary tree and returning the values of the nodes in the traversal as a vector.

The preorderTraversal function takes as input a pointer to the root node of a binary tree and returns a vector containing the values of the nodes in the tree as they are visited in a preorder traversal.

The preorder function is a helper function that performs the actual traversal and populates the vector with the node values. It does this by first adding the value of the current node to the vector, then recursively calling itself on the left and right children of the current node (in that order). If the current node is null, the function simply returns without adding anything to the vector.

The preorder function is private, meaning that it can only be called from within the Solution class. This is useful because it means that the implementation details of the traversal are hidden from users of the Solution class, who can simply call preorderTraversal to get the desired result.