Create countDigits.cpp
This code defines a class called Solution that has a single public method called evenlyDivides. This method takes an integer N as input and returns the number of digits in N that evenly divide N.

The evenlyDivides method first initializes a counter variable cnt to 0. It then stores a copy of N in a variable called n.

Next, the method enters a loop that continues as long as n is greater than 0. Inside the loop, it extracts the rightmost digit of n by taking the remainder of n divided by 10 (using the modulus operator %). It then checks whether this digit is not 0 (to avoid division by 0 errors) and whether N is evenly divisible by this digit (using the modulus operator % again). If both conditions are true, the method increments the counter.

After extracting the rightmost digit of n, the method divides n by 10 to remove the rightmost digit. This process continues until n becomes 0, at which point the loop ends.

Finally, the method returns the final value of the counter, which is the number of digits in N that evenly divide N.

The main function reads in the number of test cases t, and then enters a loop that iterates t times. For each iteration, it reads in an integer N and creates an instance of the Solution class. It then calls the evenlyDivides method on this instance and prints the returned result.