Create armstrongNum.cpp
The code first reads in the number of test cases t. Then, for each test case, it reads in a number n and calls the armstrongNumber function on it, printing the result.

The armstrongNumber function is a member function of the Solution class. It takes in a single integer n as a parameter and returns a string indicating whether n is an Armstrong number or not.

An Armstrong number is a number that is equal to the sum of the cubes of its digits. For example, 371 is an Armstrong number because 3^3 + 7^3 + 1^3 = 371.

To determine whether a given number n is an Armstrong number, the armstrongNumber function initializes a variable sum to 0 and creates a copy of n called dup. It then enters a loop that will run as long as dup is greater than 0. Inside the loop, it obtains the last digit of dup using the modulo operator (%) and adds the cube of this digit to sum. It then divides dup by 10 to get rid of the last digit.

After the loop ends, the function checks whether sum is equal to n. If it is, the function returns "Yes". Otherwise, it returns "No".

The main function handles input and output and calls the armstrongNumber function on each test case.