Create sumOfAllDivisor.cpp
The code first includes the bits/stdc++.h header file and uses the std namespace. This header file includes several standard C++ libraries, and the using namespace std statement allows us to use elements from these libraries without specifying the namespace.

Next, the Solution class is defined, which contains a function called sumOfDivisors. This function takes an integer N as input and returns the sum of all the divisors of N.

The function begins by defining a variable sum which will be used to store the sum of the divisors. It is initialized to 0. Then, a for loop iterates over all numbers from 1 to N. For each iteration of the loop, the code checks if N is divisible by the current loop variable i. If it is, it adds i to the sum.

Finally, after the loop has completed, the function returns the value of sum.

The main function then handles taking input and calling the sumOfDivisors function. It first reads in an integer t, which represents the number of test cases. Then, it enters a while loop which will run t times. For each iteration of the loop, the code reads in an integer N and creates an object of the Solution class called ob. It then calls the sumOfDivisors function on this object, passing N as an argument, and stores the result in a variable ans. The value of ans is then printed to the console.

After the loop has completed, the main function returns 0.