Create maxNoOfJumps.cpp
The minJumps function takes in an integer array arr and an integer n representing the number of elements in the array. It returns an integer representing the minimum number of jumps required to reach the end of the array.

The function first checks if the first element of the array is 0 and there are more than 1 element in the array. If this is the case, it returns -1, as it is not possible to jump forward from the first element if it is 0. If there is only 1 element in the array, the function returns 0, since no jump is required.

Next, the function initializes the maximum reachable index to 0 and the number of jumps to 0, and the current index to 0. It then iterates through the array. For each iteration, it updates the maximum reachable index by taking the maximum of the current maximum reachable index and the current index plus the element at the current index. This ensures that the maximum reachable index is always the maximum index that can be reached from the current position.

If the current index is equal to the current maximum reachable index, the number of jumps is incremented and the current index is updated to the new maximum reachable index. This is because, at this point, we have reached the maximum index that can be reached from the previous position, and a jump is required to continue moving forward.

Finally, the function checks if the current index is less than the last index of the array. If this is the case, it returns -1, as it is not possible to reach the end of the array. Otherwise, it returns the number of jumps.