Create reverseInteger.cpp
The function reverse(int x) takes an integer as input and returns its reverse. The reverse of a number is obtained by reading it from right to left and not left to right, for example, the reverse of 123 is 321.

The function begins by initializing a variable ans to store the reversed number. It also creates a copy of the input number called num. This is done because we want to modify the value of num (by removing the last digit) in the loop, but we want to keep the original value of x for later use.

Next, the function enters a while loop that continues until all the digits of num have been processed. Inside the loop, the function first checks if the current value of ans will cause an overflow or underflow. This is done by checking if ans is greater than INT_MAX / 10 or less than INT_MIN / 10. INT_MAX and INT_MIN are preprocessor macros that represent the maximum and minimum values that can be stored in an int data type. If ans is outside of this range, it means that the reversed number will not fit in an int and we return 0 as per the problem statement.

If ans is within the acceptable range, we proceed to add the next digit to it. We do this by first calculating the last digit of num using the modulus operator (%). The modulus operator returns the remainder of a division, so num % 10 will give us the last digit of num. We then add this digit to ans by multiplying ans by 10 and adding the last digit. This is done because when we reverse a number, each new digit we add goes in the ones place, so we need to multiply the current value of ans by 10 to make room for the new digit.

Finally, we remove the last digit of num by dividing it by 10 using integer division (/). This will remove the last digit because integer division discards the remainder.

After the loop finishes, the function returns the reversed number stored in ans.