Create recursion1.cpp
This code is designed to print numbers from 1 to N in increasing order, where N is an input provided to the program. The core of the program is a class called "Solution", which has a single member function called "printNos(int N)".

When the "printNos(int N)" function is called, it in turns calls a helper function called "recursion(int n, int b)" with the arguments 1 and N. The recursion function is a recursive function.

The recursion function has two inputs, the current number 'n' and the end value 'b'. The base case for the recursion is when the current number 'n' is equal to the end value 'b' + 1, at this point the function will return nothing.

Otherwise, the function will first print the current number 'n' and then increment it by 1. Then, it will make a recursive call to itself with the incremented value of 'n' as the new current number.

This process will repeat until the current number 'n' becomes equal to the end value 'b' + 1, at which point the recursion will end and all the numbers from 1 to N will have been printed in increasing order.

The driver code first reads in the number of test cases 'T' and then, for each test case, it reads in the end value 'N' for the range of numbers to be printed, and it calls the "printNos(int N)" function passing 'N' as the argument and printing a new line after each test case.