Create namesRecursion.cpp
The code is a simple implementation of a solution that uses recursion to print the string "GFG" a certain number of times. The code defines a Solution class with a single function printGfg(N). This function takes an input parameter N and uses a recursive function call to print "GFG" N times. It uses a counter variable cnt to keep track of the number of times "GFG" has been printed so far, and a base case that ends the recursion when cnt reaches N.

In the main() function, the code takes input for the number of test cases, T, and then runs a loop that calls the printGfg(N) function T times, passing in the input value of N each time. This way, the user can input a number T to test the function T number of times, each time with a different input value N.

At each function call, function will print GFG and will increment counter and then will call itself with the same input(N), and the recursion will happen until the count reaches the value of N. Then the program will end and will return to the previous function call and the process will repeat until the end of the main loop

It uses recursion, in which the function calls itself multiple times with the same input to achieve the desired output, in this case it will call printGfg(N) function N times.