Create minTimeToCollectAllApples.cpp
This code is solving the problem of finding the minimum time required to collect all the apples in a tree. The tree is represented by an adjacency list, which is an array of vectors where each index represents a node in the tree and the elements in the vector at that index represent the nodes that are connected to that node by an edge.

The boolean array hasApple represents if a particular node in the tree has an apple or not.

The function minTime(int n, vector<vector<int>>& edges, vector<bool>& hasApple) is the main function that is called to solve the problem. It first initializes the adjacency list by building it using the edges provided in the input.

Then it calls the dfs(vector<int>adj[],vector<bool> &hasApple,int node, int root) function which performs a depth-first search on the tree. This function takes four arguments: the adjacency list of the tree, the boolean array representing if a node has an apple or not, the current node being visited and the parent node of the current node.

The dfs function starts by initializing a variable dist to keep track of the total distance needed to collect all the apples. It then iterates through all the adjacent nodes of the current node using a for loop. For each adjacent node, it checks if the adjacent node is not the parent node of the current node, and if so, it recursively calls the dfs function on that adjacent node, passing the adjacent node as the current node and the current node as the parent node.

The dfs function then checks if the adjacent node has an apple by checking the boolean array hasApple. If the adjacent node has an apple, it adds the distance needed to collect that apple, which is 2 (1 to get to the adjacent node and 1 to get back), to the total distance dist. It also marks the current node as having an apple by setting the corresponding element in the boolean array hasApple to true.

After iterating through all the adjacent nodes, the dfs function returns the total distance needed to collect all the apples in the subtree rooted at the current node. The minTime function then returns the total distance returned by the dfs function.

Overall, the code uses a depth-first search algorithm to traverse the tree and collect all the apples. The time complexity of this algorithm is O(N) where N is the number of nodes in the tree, because each node is visited once.