Create freqOfLimitedRange.cpp
This code defines a C++ class called "Solution" that contains a single function called "frequencyCount". This function takes in three parameters: a reference to a vector of integers called "arr", an integer "N", and an integer "P".

The function first declares an unordered_map called "mpp" which is used to store the frequency count of all elements in the input array "arr". Then it iterates through the input array and for each element, it increments the value stored in the unordered_map at the key of that element.

After that, the function iterates through all elements from 1 to N. For each element, it checks if that element is present in the unordered_map, if it is present then it assigns the frequency count of that element from the unordered_map to the corresponding position in the input array and if it is not present then it assigns 0 to the corresponding position in the input array.

This function essentially counts the frequency of all elements from 1 to N in the array and stores the result in the input array.