Create bitManipulation.cpp
The bitManipulation function takes in two arguments: an integer num, and an integer i. The function first decrements the value of i by 1, as the bit numbering in this implementation starts from 0 instead of 1.

The function then checks the value of the i-th bit of num. It does this by right-shifting num by i bits, and then taking the modulus of the result with 2. If the result is non-zero, it means that the i-th bit of num is set (has a value of 1), and the function prints "1" followed by a space. If the result is zero, it means that the i-th bit of num is not set (has a value of 0), and the function prints "0" followed by a space.

The function then sets the i-th bit of num by performing a bitwise OR operation with the value of 1 shifted left by i bits. This results in the i-th bit of num being set to 1, while leaving the other bits unchanged. The function then prints the result of this operation followed by a space.

The function then clears the i-th bit of num by performing a bitwise AND operation with the bitwise NOT of the value of 1 shifted left by i bits. This results in the i-th bit of num being set to 0, while leaving the other bits unchanged. The function then prints the result of this operation.