Create KthElementSetOrNot.cpp
The code is a C++ class called Solution which contains a function called checkKthBit(). This function takes in two parameters, an integer n and an integer k.

The function checks if the kth bit of the binary representation of n is set or not.

It uses the bitwise AND operator & to check the kth bit. It first shifts the number 1 by k positions to the left using the left shift operator <<. This results in a number with only the kth bit set (1) and all other bits set to 0.

Then it uses the AND operator & to perform a bitwise AND operation between the kth bit of n and the number obtained by left shifting 1.

If the kth bit of n is also set, the result of this operation will be non-zero and the function returns true. Otherwise, the result will be zero and the function returns false.

The function returns a boolean value, true if the kth bit is set, false otherwise.