Create powOf2Bit.cpp
The function isPowerofTwo(long long n) takes a long long number as input and checks whether it is a power of 2 or not. The function first checks if the input number is zero and returns false if it is. This is because 0 is not a power of 2, as any power of 2 must be greater than 0.

Then, it uses the bitwise AND operator to check if the number is a power of 2. The bitwise AND operator compares each bit of the first operand to the corresponding bit of the second operand. If both bits are 1, the corresponding result bit is set to 1. Otherwise, the corresponding result bit is set to 0. In this case, the first operand is n and the second operand is n-1.

If the number is a power of 2, it will only have one bit set to 1 and all the others will be 0. For example, the binary representation of 4 (which is a power of 2) is 100, and the binary representation of 3 (which is one less than 4) is 011. When we perform the bitwise AND operation between 4 and 3, it will be:

   100
&  011
   ---
   000
As you can see, all the bits in the result are 0, so the result of the AND operation is 0.

Therefore, the expression n & (n-1) will be 0 for a power of 2. The function returns true if the number is a power of 2, and false otherwise.